+++ src/core/document/json.pl
@@ -59,13 +59,20 @@
 encode_id_fragment(Elt, Encoded) :-
     ground(Elt),
     !,
-    format(atom(Fragment), '~w', [Elt]),
-    uri_encoded(segment, Fragment, Part_Encoded),
-    re_replace('\\+'/g, '%2B', Part_Encoded, Encoded).
+    (   Elt = optional(none)
+    ->  Encoded = "+none+"
+    ;   Elt = list(List)
+    ->  maplist(encode_id_fragment, List, List_Encoded),
+        merge_separator_split(Encoded, '++', List_Encoded)
+    ;   format(atom(Fragment), '~w', [Elt]),
+        uri_encoded(segment, Fragment, Part_Encoded),
+        re_replace('\\+'/g, '%2B', Part_Encoded, Encoded)).
 encode_id_fragment(Fragment, Encoded) :-
     ground(Encoded),
     !,
-    uri_encoded(segment, Fragment, Encoded).
+    (   Encoded = "+"
+    ->  Fragment = optional(none)
+    ;   uri_encoded(segment, Fragment, Encoded)).
 encode_id_fragment(_, _) :-
     throw(error(instantiation_error, _)).
 
@@ -191,14 +198,21 @@ get_path_value(Elaborated,[I|Path],Value) :-
     nth0(I, Sorted, Elt),
     get_path_value(Elt, Path, Value).
 
-get_field_values(JSON,Context,Fields,Values) :-
+get_field_values(JSON,DB,Context,Fields,Values) :-
     findall(
         Value,
         (   member(Field,Fields),
             prefix_expand_schema(Field,Context,Field_Ex),
             (   get_dict(Field_Ex,JSON,Value)
             ->  true
-            ;   throw(error(missing_key(Field,JSON),_))
+            %   key field not found! We'll have to find out if that's allowable or not
+            ;   get_dict('@type',JSON,Type),
+                prefix_expand_schema(Type, Context, Type_Ex),
+                prefix_expand_schema(Field, Context, Field_Ex),
+                class_predicate_type(DB, Type_Ex, Field_Ex, Field_Type),
+                memberchk(Field_Type, [optional(_), set(_), array(_)])
+            ->  Value = optional(none)
+            ;   throw(error(key_missing_required_field(Field),_))
             )
         ),
         Values).
@@ -208,21 +222,35 @@ raw(JValue,Value) :-
     !,
     (   get_dict('@type', JValue, "@id")
     ->  get_dict('@id', JValue, Value)
-    ;   get_dict('@value',JValue,Value)
+    ;   get_dict('@container', JValue, Container_Type)
+    ->  get_dict('@value', JValue, V),
+        (   [] = V
+        ->  Value = optional(none)
+        ;   [Single_Value] = V
+        ->  raw(Single_Value, Value)
+        ;   \+ is_list(V)
+        ->  Value = V
+        ;   (   Container_Type = "@set"
+            ->  sort(V, V_1)
+            ;   V_1 = V),
+            maplist(raw, V_1, V_Raw),
+            Value = list(V_Raw))
+    ;   get_dict('@value', JValue, Value)
     ).
 raw(JValue,JValue).
 
-idgen_lexical(Base,Values,ID) :-
+idgen_suffix(Values, Suffix) :-
     maplist(raw,Values,Raw),
     maplist(encode_id_fragment,Raw,Encoded),
-    merge_separator_split(Suffix, '+', Encoded),
+    merge_separator_split(Suffix, '+', Encoded).
+
+idgen_lexical(Base,Values,ID) :-
+    idgen_suffix(Values, Suffix),
     format(string(ID), '~w~w', [Base,Suffix]).
 
 idgen_hash(Base,Values,ID) :-
-    maplist(raw,Values,Raw),
-    maplist(encode_id_fragment,Raw,Encoded),
-    merge_separator_split(String, '+', Encoded),
-    crypto_data_hash(String, Hash, [algorithm(sha256)]),
+    idgen_suffix(Values, Suffix),
+    crypto_data_hash(Suffix, Hash, [algorithm(sha256)]),
     format(string(ID), "~w~w", [Base,Hash]).
 
 idgen_path_values_hash(Base,Path,ID) :-
@@ -282,11 +310,11 @@ json_idgen(JSON,DB,Context,Path,ID_Ex) :-
     get_dict('@type',JSON,Type),
     key_descriptor(DB,Context,Type,Descriptor),
     (   Descriptor = lexical(Base,Fields)
-    ->  get_field_values(JSON, Context, Fields, Values),
+    ->  get_field_values(JSON, DB, Context, Fields, Values),
         path_component([type(Base)|Path], Context, [Path_Base]),
         idgen_lexical(Path_Base,Values,ID)
     ;   Descriptor = hash(Base,Fields)
-    ->  get_field_values(JSON, Context, Fields, Values),
+    ->  get_field_values(JSON, DB, Context, Fields, Values),
         path_component([type(Base)|Path], Context, [Path_Base]),
         idgen_hash(Path_Base,Values,ID)
     ;   Descriptor = value_hash(Base)
@@ -575,10 +603,6 @@ context_value_expand(_,_,[],json{},[]) :-
     !.
 context_value_expand(_,_,null,_,null) :-
     !.
-context_value_expand(_,_,true,_,{"@type" : "xsd:boolean", "@value" : true}) :-
-    !.
-context_value_expand(_,_,false,_,{"@type" : "xsd:boolean", "@value" : false}) :-
-    !.
 context_value_expand(DB,Context,Value,Expansion,V) :-
     get_dict('@container', Expansion, _),
     !,
@@ -766,7 +790,7 @@ is_type_enum(JSON) :-
 
 context_triple(JSON,Triple) :-
     context_elaborate(JSON,Elaborated),
-    expand(Elaborated,JSON{
+    expand(Elaborated,json{
                           sys:'http://terminusdb.com/schema/sys#',
                           xsd:'http://www.w3.org/2001/XMLSchema#',
                           xdd:'http://terminusdb.com/schema/xdd#'
@@ -897,6 +921,13 @@ json_schema_elaborate_key(V,_,json{ '@type' : Type}) :-
     get_dict('@type', V, Random),
     expand_match_system(Random, 'Random', Type),
     !.
+json_schema_elaborate_key(V,_,_) :-
+    get_dict('@type', V, Type),
+    !,
+    throw(error(document_key_type_unknown(Type), _)).
+json_schema_elaborate_key(V,_,_) :-
+    atom_json_dict(Atom, V, []),
+    throw(error(document_key_type_missing(Atom), _)).
 
 json_schema_elaborate_property_documentation(Context, Path, Dict, Out) :-
     global_prefix_expand(sys:'PropertyDocumentation',Property_Ex),
@@ -1018,8 +1049,15 @@ json_schema_predicate_value(P,V,Context,_,Prop,json{'@type' : "@id",
     prefix_expand_schema(P,Context,Prop),
     prefix_expand_schema(V,Context,VEx).
 
-json_schema_elaborate(JSON,Context,_,Elaborated) :-
-    is_type_enum(JSON),
+json_schema_elaborate(JSON,Context,Path,Elaborated) :-
+    get_dict('@type', JSON, Type),
+    compress_system_uri(Type,Context,Type_Min),
+    do_or_die(
+        json_schema_elaborate_(Type_Min,JSON,Context,Path,Elaborated),
+        error(schema_type_unknown(Type_Min),_)
+    ).
+
+json_schema_elaborate_('Enum',JSON,Context,_,Elaborated) :-
     !,
     get_dict('@id', JSON, ID),
     prefix_expand_schema(ID,Context,ID_Ex),
@@ -1038,15 +1076,19 @@ json_schema_elaborate(JSON,Context,_,Elaborated) :-
                               json{ '@container' : "@list",
                                     '@type' : "@id",
                                     '@value' : New_List })).
-json_schema_elaborate(JSON,Context,Old_Path,Elaborated) :-
+json_schema_elaborate_(Type,JSON,Context,Old_Path,Elaborated) :-
+    memberchk(Type,['Class','TaggedUnion',
+                    'Set','List','Optional','Array', 'Cardinality',
+                    'Foreign']),
     is_dict(JSON),
     dict_pairs(JSON,json,Pre_Pairs),
     !,
     (   is_type_family(JSON)
     ->  type_family_id(JSON,Context,Old_Path,ID),
         Pairs = ['@id'-ID|Pre_Pairs]
-    ;   Pairs = Pre_Pairs,
-        get_dict('@id',JSON,ID)
+    ;   get_dict('@id',JSON,ID)
+    ->  Pairs = Pre_Pairs
+    ;   throw(error(no_id_in_document(JSON), _))
     ),
     Path = [type(ID)|Old_Path],
     findall(
@@ -1067,24 +1109,29 @@ check_schema_document_restrictions(Elaborated) :-
     get_dict(AbsP, Elaborated, _),
     !.
 check_schema_document_restrictions(Elaborated) :-
+    global_prefix_expand(sys:key, KeyP),
+    do_or_die(get_dict(KeyP, Elaborated, Key),
+              error(subdocument_key_missing, _)),
+    % We currently don't reach the following checks because
+    % json_schema_elaborate_key throw errors for the same conditions.
+    do_or_die(get_dict('@type', Key, Key_Type_String),
+              error(subdocument_key_type_missing,_)),
+    atom_string(Key_Type, Key_Type_String),
     do_or_die(
-        (   global_prefix_expand(sys:key, KeyP),
-            get_dict(KeyP, Elaborated, Key),
-            % Is this not exhaustive?
-            (   global_prefix_expand(sys:'ValueHash',Key_Type)
-            ;   global_prefix_expand(sys:'Hash',Key_Type)
-            ;   global_prefix_expand(sys:'Lexical',Key_Type)
-            ;   global_prefix_expand(sys:'Random',Key_Type)),
-            get_dict('@type', Key, Key_Type_String),
-            atom_string(Key_Type, Key_Type_String)
-        ),
-        error(subdocument_key_type_restriction,_)).
+        % Is this exhaustive?
+        (   global_prefix_expand(sys:'ValueHash',Key_Type)
+        ;   global_prefix_expand(sys:'Hash',Key_Type)
+        ;   global_prefix_expand(sys:'Lexical',Key_Type)
+        ;   global_prefix_expand(sys:'Random',Key_Type)),
+        error(subdocument_key_type_unknown(Key_Type_String),_)).
 
 json_schema_elaborate(JSON,Context,JSON_Schema) :-
     json_schema_elaborate(JSON,Context,[],JSON_Schema).
 
 json_schema_triple(JSON,Context,Triple) :-
-    json_schema_elaborate(JSON,Context,[],JSON_Schema),
+    do_or_die(
+        json_schema_elaborate(JSON,Context,[],JSON_Schema),
+        error(unable_to_elaborate_schema_document(JSON),_)),
     json_triple_(JSON_Schema,Context,Triple).
 
 % Triple generator
@@ -1330,6 +1377,15 @@ type_id_predicate_iri_value(base_class(C),_,_,X^^T,_,_,Prefixes,_Compress,_Unfol
         compress_dict_uri(T2,Prefixes,T2C),
         V = json{ '@type' : T2C, '@value' : D}
     ).
+
+
+compress_system_uri(IRI,Prefixes,IRI_Atom) :-
+    put_dict(_{'@base' : 'http://terminusdb.com/schema/sys#'}, Prefixes, Schema_Prefixes),
+    compress_dict_uri(IRI,Schema_Prefixes,IRI_Comp),
+    atom_string(IRI_Atom, IRI_Comp),
+    !.
+compress_system_uri(IRI,_Prefixes,IRI_Atom) :-
+    atom_string(IRI_Atom, IRI).
 
 compress_schema_uri(IRI,Prefixes,IRI_Comp) :-
     (   get_dict('@schema',Prefixes,Schema),
@@ -2215,7 +2271,8 @@ replace_schema_document(Transaction, Document, Id) :-
         insert_schema_document_unsafe(Transaction, Document)
     ;   get_dict('@type', Document, "@context")
     ->  delete_schema_document(Transaction, 'terminusdb://context'),
-        insert_context_document(Transaction, Document)
+        insert_context_document(Transaction, Document),
+        Id='@context'
     ;   throw(error(no_id_in_document(Document),_))
     ).
 replace_schema_document(Query_Context, Document, Id) :-
@@ -2435,9 +2492,9 @@ test(get_field_values, []) :-
                          '@value':"1979-12-28"},
                     'http://s/name':json{'@type':'http://www.w3.org/2001/XMLSchema#string',
                                          '@value':"jane"}},
-    get_field_values(Expanded, context{'@base':"http://i/",
-                                       '@schema':"http://s/",
-                                       '@type':'http://terminusdb.com/schema/sys#Context'},
+    get_field_values(Expanded, _, context{'@base':"http://i/",
+                                          '@schema':"http://s/",
+                                          '@type':'http://terminusdb.com/schema/sys#Context'},
                      [name, birthdate],
                      [ json{'@type':'http://www.w3.org/2001/XMLSchema#string',
                             '@value':"jane"},
@@ -3044,6 +3101,231 @@ test(idgen_value_hash,
           '@type':'http://s/Task',
           'http://s/name':json{ '@type':'http://www.w3.org/2001/XMLSchema#string',
                                 '@value':"Groceries"}}.
+
+test(idgen_lexical_optional,
+     [setup((setup_temp_store(State),
+             create_db_with_empty_schema("admin", "testdb"),
+             resolve_absolute_string_descriptor("admin/testdb", Desc))),
+      cleanup(teardown_temp_store(State))]) :-
+
+    with_test_transaction(Desc,
+                          C1,
+                          insert_schema_document(
+                              C1,
+                              _{'@type': "Class",
+                                '@id': "Thing",
+                                '@key': _{'@type': "Lexical",
+                                          '@fields': ["field"]},
+                                field: _{'@type': "Optional",
+                                         '@class': "xsd:string"}})
+                          ),
+
+    with_test_transaction(Desc,
+                          C2,
+                          (   insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: "foo"},
+                                  Uri1),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing"},
+                                  Uri2)
+                          )),
+
+    Uri1 = 'http://somewhere.for.now/document/Thing/foo',
+    Uri2 = 'http://somewhere.for.now/document/Thing/+none+'.
+
+test(idgen_lexical_set,
+     [setup((setup_temp_store(State),
+             create_db_with_empty_schema("admin", "testdb"),
+             resolve_absolute_string_descriptor("admin/testdb", Desc))),
+      cleanup(teardown_temp_store(State))]) :-
+
+    with_test_transaction(Desc,
+                          C1,
+                          insert_schema_document(
+                              C1,
+                              _{'@type': "Class",
+                                '@id': "Thing",
+                                '@key': _{'@type': "Lexical",
+                                          '@fields': ["field"]},
+                                field: _{'@type': "Set",
+                                         '@class': "xsd:string"}})
+                          ),
+
+    with_test_transaction(Desc,
+                          C2,
+                          (   insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["foo", "bar"]},
+                                  Uri1),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["bar", "foo"]},
+                                  Uri2),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["quux"]},
+                                  Uri3),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: []},
+                                  Uri4),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing"},
+                                  Uri5)
+                          )),
+
+    Uri1 = 'http://somewhere.for.now/document/Thing/bar++foo',
+    Uri2 = 'http://somewhere.for.now/document/Thing/bar++foo',
+    Uri3 = 'http://somewhere.for.now/document/Thing/quux',
+    Uri4 = 'http://somewhere.for.now/document/Thing/+none+',
+    Uri5 = 'http://somewhere.for.now/document/Thing/+none+'.
+
+test(idgen_lexical_list,
+     [setup((setup_temp_store(State),
+             create_db_with_empty_schema("admin", "testdb"),
+             resolve_absolute_string_descriptor("admin/testdb", Desc))),
+      cleanup(teardown_temp_store(State))]) :-
+
+    with_test_transaction(Desc,
+                          C1,
+                          insert_schema_document(
+                              C1,
+                              _{'@type': "Class",
+                                '@id': "Thing",
+                                '@key': _{'@type': "Lexical",
+                                          '@fields': ["field"]},
+                                field: _{'@type': "List",
+                                         '@class': "xsd:string"}})
+                          ),
+
+    with_test_transaction(Desc,
+                          C2,
+                          (   insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["foo", "bar"]},
+                                  Uri1),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["bar", "foo"]},
+                                  Uri2),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["quux"]},
+                                  Uri3),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: []},
+                                  Uri4)
+                          )),
+
+    Uri1 = 'http://somewhere.for.now/document/Thing/foo++bar',
+    Uri2 = 'http://somewhere.for.now/document/Thing/bar++foo',
+    Uri3 = 'http://somewhere.for.now/document/Thing/quux',
+    Uri4 = 'http://somewhere.for.now/document/Thing/+none+'.
+
+test(idgen_lexical_array,
+     [setup((setup_temp_store(State),
+             create_db_with_empty_schema("admin", "testdb"),
+             resolve_absolute_string_descriptor("admin/testdb", Desc))),
+      cleanup(teardown_temp_store(State))]) :-
+
+    with_test_transaction(Desc,
+                          C1,
+                          insert_schema_document(
+                              C1,
+                              _{'@type': "Class",
+                                '@id': "Thing",
+                                '@key': _{'@type': "Lexical",
+                                          '@fields': ["field"]},
+                                field: _{'@type': "Array",
+                                         '@class': "xsd:string"}})
+                          ),
+
+    with_test_transaction(Desc,
+                          C2,
+                          (   insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["foo", "bar"]},
+                                  Uri1),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["bar", "foo"]},
+                                  Uri2),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["quux"]},
+                                  Uri3),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: []},
+                                  Uri4),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing"},
+                                  Uri5)
+                          )),
+
+    Uri1 = 'http://somewhere.for.now/document/Thing/foo++bar',
+    Uri2 = 'http://somewhere.for.now/document/Thing/bar++foo',
+    Uri3 = 'http://somewhere.for.now/document/Thing/quux',
+    Uri4 = 'http://somewhere.for.now/document/Thing/+none+',
+    Uri5 = 'http://somewhere.for.now/document/Thing/+none+'.
+
+test(idgen_find_collision,
+     [setup((setup_temp_store(State),
+             create_db_with_empty_schema("admin", "testdb"),
+             resolve_absolute_string_descriptor("admin/testdb", Desc))),
+      cleanup(teardown_temp_store(State))]) :-
+
+    with_test_transaction(Desc,
+                          C1,
+                          insert_schema_document(
+                              C1,
+                              _{'@type': "Class",
+                                '@id': "Thing",
+                                '@key': _{'@type': "Lexical",
+                                          '@fields': ["field"]},
+                                field: _{'@type': "Array",
+                                         '@class': "xsd:string"}})
+                          ),
+
+    with_test_transaction(Desc,
+                          C2,
+                          (   insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["", "none", ""]},
+                                  Uri1),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: []},
+                                  Uri2),
+                              insert_document(
+                                  C2,
+                                  _{'@type': "Thing",
+                                    field: ["none"]},
+                                  Uri3)
+                          )),
+    Uri1 = 'http://somewhere.for.now/document/Thing/++none++',
+    Uri2 = 'http://somewhere.for.now/document/Thing/+none+',
+    Uri3 = 'http://somewhere.for.now/document/Thing/none'.
 
 test(idgen_random,
      [
@@ -4162,8 +4444,7 @@ test(partial_document_elaborate_list_without_required,
          cleanup(
              teardown_temp_store(State)
          ),
-         error(missing_key('http://s/name',
-                           _),
+         error(key_missing_required_field('http://s/name'),
                _)
      ]) :-
 
@@ -5196,6 +5477,35 @@ test(elaborate_null,
           'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
           'http://somewhere.for.now/schema#Doc')
     ].
+
+schema10('
+{ "@type" : "@context",
+  "@base" : "http://i/",
+  "@schema" : "http://s/" }
+{ "@type" : "Class",
+  "@id" : "Boolean",
+  "b" : "xsd:boolean" }
+').
+
+test(boolean_in_boolean_field,
+     [
+         setup(
+             (   setup_temp_store(State),
+                 create_db_with_empty_schema("admin", "foo"),
+                 resolve_absolute_string_descriptor("admin/foo", Desc),
+                 write_schema(schema10, Desc)
+             )),
+         cleanup(
+             teardown_temp_store(State)
+         )
+     ]) :-
+    Document = _{ '@type': "Boolean", b: true },
+    create_context(Desc, _{ author: "a", message: "m" }, Context),
+    with_transaction(
+        Context,
+        insert_document(Context, Document, _Id),
+        _
+    ).
 
 :- end_tests(json).